Fix OpenGL code to work on Intel Linux

+++ Render.cpp
@@ -596,9 +596,9 @@ void Render::render(float64 t, float64 time)
       shader->set_uniform((s("lights[") + ts(i) + s("].cone_angle")).c_str(),
                           entity.lights.lights[i].cone_angle);
       shader->set_uniform((s("lights[") + ts(i) + s("].type")).c_str(),
-                          (uint32)entity.lights.lights[i].type);
+                          (int32)entity.lights.lights[i].type);
     }
-    shader->set_uniform("number_of_lights", entity.lights.count);
+    shader->set_uniform("number_of_lights", (int32)entity.lights.count);
 
     //// verify sizes of data, mat4 floats
     ASSERT(entity.Model_Matrices.size() > 0);
